增加学院管理功能的异常处理，并在前台页面上使用Alert将其输出。

+++ src/main/java/com/houseyoung/student/service/impl/SchoolServiceImpl.java
@@ -4,6 +4,7 @@ import com.houseyoung.student.entity.School;
 import com.houseyoung.student.mapper.SchoolMapper;
 import com.houseyoung.student.service.SchoolService;
 import org.apache.ibatis.annotations.Param;
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,27 +19,61 @@ public class SchoolServiceImpl implements SchoolService{
     private SchoolMapper schoolMapper;
 
     @Override
-    public List<School> listSchool(@Param("keywords") String keywords){
-        return schoolMapper.listSchool(keywords);
+    public List<School> listSchool(@Param("keywords") String keywords) throws Exception{
+        try {
+            return schoolMapper.listSchool(keywords);
+        } catch (Exception e){
+            throw new Exception("系统内部异常");
+        }
     }
 
     @Override
-    public void insert(@Param("school") School school){
-        schoolMapper.insert(school);
+    public void insert(@Param("school") School school) throws Exception{
+        try {
+            if (school.getName() == null || school.getName() == ""){
+                throw new Exception("学院名不可为空");
+            }
+            //TODO: 判断是否重名
+            schoolMapper.insert(school);
+        } catch (Exception e){
+            throw e;
+        }
     }
 
     @Override
-    public void delete(@Param("id") int id)  {
-        schoolMapper.delete(id);
+    public void delete(@Param("id") int id) throws Exception{
+        try {
+            if (id <= 0){
+                throw new Exception("ID不合法");
+            }
+            schoolMapper.delete(id);
+        } catch (Exception e){
+            throw e;
+        }
     }
 
     @Override
-    public School queryById(@Param("id") int id) {
-        return schoolMapper.queryById(id);
+    public School queryById(@Param("id") int id) throws Exception{
+        try {
+            if (id <= 0){
+                throw new Exception("ID不合法");
+            }
+            return schoolMapper.queryById(id);
+        } catch (Exception e){
+            throw e;
+        }
     }
 
     @Override
-    public void update(@Param("school") School school){
-        schoolMapper.update(school);
+    public void update(@Param("school") School school) throws Exception{
+        try {
+            if (school.getName() == null || school.getName() == ""){
+                throw new Exception("学院名不可为空");
+            }
+            //TODO: 判断是否重名
+            schoolMapper.update(school);
+        } catch (Exception e){
+            throw e;
+        }
     }
 }
